services/api-gateway/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
httpx==0.23.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0 